# ✨ Add `constructorHandlers` option