--- a/index.js
+++ b/index.js
@@ -11,8 +11,14 @@ function copyBuffer (cur) {
 
 function rfdc (opts) {
   opts = opts || {}
-
   if (opts.circles) return rfdcCircles(opts)
+
+  const constructorHandlers = new Map([
+    [Date, (o) => new Date(o)],
+    [Map, (o, fn) => new Map(cloneArray(Array.from(o), fn))],
+    [Set, (o, fn) => new Set(cloneArray(Array.from(o), fn))]
+  ].concat(opts.constructorHandlers || []))
+
   return opts.proto ? cloneProto : clone
 
   function cloneArray (a, fn) {
@@ -23,8 +29,8 @@ function rfdc (opts) {
       const cur = a[k]
       if (typeof cur !== 'object' || cur === null) {
         a2[k] = cur
-      } else if (cur instanceof Date) {
-        a2[k] = new Date(cur)
+      } else if (constructorHandlers.has(cur.constructor)) {
+        a2[k] = constructorHandlers.get(cur.constructor)(cur, fn)
       } else if (ArrayBuffer.isView(cur)) {
         a2[k] = copyBuffer(cur)
       } else {
@@ -36,22 +42,18 @@ function rfdc (opts) {
 
   function clone (o) {
     if (typeof o !== 'object' || o === null) return o
-    if (o instanceof Date) return new Date(o)
     if (Array.isArray(o)) return cloneArray(o, clone)
-    if (o instanceof Map) return new Map(cloneArray(Array.from(o), clone))
-    if (o instanceof Set) return new Set(cloneArray(Array.from(o), clone))
+    if (constructorHandlers.has(o.constructor)) {
+      return constructorHandlers.get(o.constructor)(o, clone)
+    }
     const o2 = {}
     for (const k in o) {
       if (Object.hasOwnProperty.call(o, k) === false) continue
       const cur = o[k]
       if (typeof cur !== 'object' || cur === null) {
         o2[k] = cur
-      } else if (cur instanceof Date) {
-        o2[k] = new Date(cur)
-      } else if (cur instanceof Map) {
-        o2[k] = new Map(cloneArray(Array.from(cur), clone))
-      } else if (cur instanceof Set) {
-        o2[k] = new Set(cloneArray(Array.from(cur), clone))
+      } else if (constructorHandlers.has(cur.constructor)) {
+        o2[k] = constructorHandlers.get(cur.constructor)(cur, clone)
       } else if (ArrayBuffer.isView(cur)) {
         o2[k] = copyBuffer(cur)
       } else {
@@ -63,21 +65,17 @@ function rfdc (opts) {
 
   function cloneProto (o) {
     if (typeof o !== 'object' || o === null) return o
-    if (o instanceof Date) return new Date(o)
     if (Array.isArray(o)) return cloneArray(o, cloneProto)
-    if (o instanceof Map) return new Map(cloneArray(Array.from(o), cloneProto))
-    if (o instanceof Set) return new Set(cloneArray(Array.from(o), cloneProto))
+    if (constructorHandlers.has(o.constructor)) {
+      return constructorHandlers.get(o.constructor)(o, cloneProto)
+    }
     const o2 = {}
     for (const k in o) {
       const cur = o[k]
       if (typeof cur !== 'object' || cur === null) {
         o2[k] = cur
-      } else if (cur instanceof Date) {
-        o2[k] = new Date(cur)
-      } else if (cur instanceof Map) {
-        o2[k] = new Map(cloneArray(Array.from(cur), cloneProto))
-      } else if (cur instanceof Set) {
-        o2[k] = new Set(cloneArray(Array.from(cur), cloneProto))
+      } else if (constructorHandlers.has(cur.constructor)) {
+        o2[k] = constructorHandlers.get(cur.constructor)(cur, cloneProto)
       } else if (ArrayBuffer.isView(cur)) {
         o2[k] = copyBuffer(cur)
       } else {
@@ -92,6 +90,12 @@ function rfdcCircles (opts) {
   const refs = []
   const refsNew = []
 
+  const constructorHandlers = new Map([
+    [Date, (o) => new Date(o)],
+    [Map, (o, fn) => new Map(cloneArray(Array.from(o), fn))],
+    [Set, (o, fn) => new Set(cloneArray(Array.from(o), fn))]
+  ].concat(opts.constructorHandlers || []))
+
   return opts.proto ? cloneProto : clone
 
   function cloneArray (a, fn) {
@@ -102,8 +106,8 @@ function rfdcCircles (opts) {
       const cur = a[k]
       if (typeof cur !== 'object' || cur === null) {
         a2[k] = cur
-      } else if (cur instanceof Date) {
-        a2[k] = new Date(cur)
+      } else if (constructorHandlers.has(cur.constructor)) {
+        a2[k] = constructorHandlers.get(cur.constructor)(cur, fn)
       } else if (ArrayBuffer.isView(cur)) {
         a2[k] = copyBuffer(cur)
       } else {
@@ -120,10 +124,10 @@ function rfdcCircles (opts) {
 
   function clone (o) {
     if (typeof o !== 'object' || o === null) return o
-    if (o instanceof Date) return new Date(o)
     if (Array.isArray(o)) return cloneArray(o, clone)
-    if (o instanceof Map) return new Map(cloneArray(Array.from(o), clone))
-    if (o instanceof Set) return new Set(cloneArray(Array.from(o), clone))
+    if (constructorHandlers.has(o.constructor)) {
+      return constructorHandlers.get(o.constructor)(o, clone)
+    }
     const o2 = {}
     refs.push(o)
     refsNew.push(o2)
@@ -132,12 +136,8 @@ function rfdcCircles (opts) {
       const cur = o[k]
       if (typeof cur !== 'object' || cur === null) {
         o2[k] = cur
-      } else if (cur instanceof Date) {
-        o2[k] = new Date(cur)
-      } else if (cur instanceof Map) {
-        o2[k] = new Map(cloneArray(Array.from(cur), clone))
-      } else if (cur instanceof Set) {
-        o2[k] = new Set(cloneArray(Array.from(cur), clone))
+      } else if (constructorHandlers.has(cur.constructor)) {
+        o2[k] = constructorHandlers.get(cur.constructor)(cur, clone)
       } else if (ArrayBuffer.isView(cur)) {
         o2[k] = copyBuffer(cur)
       } else {
@@ -156,10 +156,10 @@ function rfdcCircles (opts) {
 
   function cloneProto (o) {
     if (typeof o !== 'object' || o === null) return o
-    if (o instanceof Date) return new Date(o)
     if (Array.isArray(o)) return cloneArray(o, cloneProto)
-    if (o instanceof Map) return new Map(cloneArray(Array.from(o), cloneProto))
-    if (o instanceof Set) return new Set(cloneArray(Array.from(o), cloneProto))
+    if (constructorHandlers.has(o.constructor)) {
+      return constructorHandlers.get(o.constructor)(o, cloneProto)
+    }
     const o2 = {}
     refs.push(o)
     refsNew.push(o2)
@@ -167,12 +167,8 @@ function rfdcCircles (opts) {
       const cur = o[k]
       if (typeof cur !== 'object' || cur === null) {
         o2[k] = cur
-      } else if (cur instanceof Date) {
-        o2[k] = new Date(cur)
-      } else if (cur instanceof Map) {
-        o2[k] = new Map(cloneArray(Array.from(cur), cloneProto))
-      } else if (cur instanceof Set) {
-        o2[k] = new Set(cloneArray(Array.from(cur), cloneProto))
+      } else if (constructorHandlers.has(cur.constructor)) {
+        o2[k] = constructorHandlers.get(cur.constructor)(cur, cloneProto)
       } else if (ArrayBuffer.isView(cur)) {
         o2[k] = copyBuffer(cur)
       } else {